requirements.txt
rerun-sdk
mujoco
colorama
tqdm
cffi
transformations
numpy==1.23  #the ROS1 point_cloud python package is incompatible with 1.24